Added comments
Not everything in the code is comented but the main things that may need some clarification are

+++ Practica3/resolvers/BorrarCliente.ts
@@ -5,14 +5,7 @@ import { HipotecaModel } from '../collections/Hipoteca.ts';
 
 export const BorrarCliente = async(req: Request, res: Response) => {
     const idCliente = req.params.id;
-
-    const hipotecasCliente = await HipotecaModel.find({idCliente: idCliente}).exec();
     
-    if(hipotecasCliente.length !== 0){
-        res.status(400).send(`El cliente aun tiene hipotecas a su nombre, por favor complete sus pagos para poder eliminar el cliente`);
-        return;
-    }
-
     const cliente = await ClienteModel.findOne().where("_id").equals(idCliente).exec();
 
     if(!cliente){
@@ -20,10 +13,12 @@ export const BorrarCliente = async(req: Request, res: Response) => {
         return;
     }
 
-    if(cliente.idGestor !== ""){
-        const gestor = await GestorModel.findOne().where("_id").equals(cliente.idGestor).exec();
-        const newNumeroClientes = gestor.numeroClientes - 1;
-        await GestorModel.findOneAndUpdate({_id: cliente.idGestor},{numeroClientes:newNumeroClientes}).exec();
+    const hipotecasCliente = await HipotecaModel.find({idCliente: idCliente}).exec(); //Cogemos todas las hipotecas que corresponden al cliente
+    
+    //No se puede eliminar un cliente si aun le quedan hipotecas por pagar
+    if(hipotecasCliente.length !== 0){
+        res.status(400).send(`El cliente aun tiene hipotecas a su nombre, por favor complete sus pagos para poder eliminar el cliente`);
+        return;
     }
 
     const deletedElements = await ClienteModel.deleteOne().where("_id").equals(idCliente).exec();
@@ -32,5 +27,12 @@ export const BorrarCliente = async(req: Request, res: Response) => {
         return;
     }
 
+    //Si el cliente tiene gestor lo actualizaremos para indicarle que tiene un cliente menos
+    if(cliente.idGestor !== ""){
+        const gestor = await GestorModel.findOne().where("_id").equals(cliente.idGestor).exec();
+        const newNumeroClientes = gestor.numeroClientes - 1;//aunque sale en rojo, el gestor siempre va a existir por como organizamos el programa y BBDD
+        await GestorModel.findOneAndUpdate({_id: cliente.idGestor},{numeroClientes:newNumeroClientes}).exec();
+    }    
+    
     res.status(200).send(`Cliente eliminado`);
 }
